chore: ⚙️ Refactor env vars
