+++ backend/src/core/config/db.config.ts
@@ -1,0 +1,11 @@
+import { config } from 'dotenv';
+
+config();
+
+export const DbConfig = {
+  HOST: process.env.DB_HOST || 'localhost',
+  PORT: parseInt(process.env.DB_PORT || '5432', 10),
+  USERNAME: process.env.DB_USER || 'postgres',
+  PASSWORD: process.env.DB_PASSWORD || 'postgres',
+  DATABASE: process.env.DB_DATABASE || 'backend_db',
+};
