+++ backend/src/core/config/index.ts
@@ -1,0 +1,2 @@
+export { AppConfig } from './app.config';
+export { DbConfig } from './db.config';
